check if user is root

+++ latest/docker-entrypoint.d/40-install-plugins.sh
@@ -1,21 +1,23 @@
 #!/bin/sh
 
-until nc -z -v -w30 "$MARIADB_HOST" 3306; do
-    echo "Waiting for MySQL connection..."
-    sleep 2
-done
-
-if [ -n "$ITSMNG_PLUGINS" ]; then
-    echo "Installation des plugins ITSM-NG..."
-
-    PACKAGES=""
-
-    for plugin in $ITSMNG_PLUGINS; do
-        PACKAGES="$PACKAGES itsm-ng-plugin-$plugin"
+if [ "$EUID" -eq 0 ]; then
+    until nc -z -v -w30 "$MARIADB_HOST" 3306; do
+        echo "Waiting for MySQL connection..."
+        sleep 2
     done
 
-    apt-get update
-    apt-get install -y $PACKAGES
+    if [ -n "$ITSMNG_PLUGINS" ]; then
+        echo "Installation des plugins ITSM-NG..."
 
-    rm -rf /var/lib/apt/lists/*
+        PACKAGES=""
+
+        for plugin in $ITSMNG_PLUGINS; do
+            PACKAGES="$PACKAGES itsm-ng-plugin-$plugin"
+        done
+
+        apt-get update
+        apt-get install -y $PACKAGES
+
+        rm -rf /var/lib/apt/lists/*
+    fi
 fi
